+++ docs/changelog.md
@@ -21,8 +21,7 @@
   - added `IsObjectProperty` helper class
 * Added support for a formal approach to writing IsXXX() classes
   - added `Check` interface
-  - added `ListCheck` interface
-  - added `ListCheckHelper` trait
+  - added `CheckList` base class
 * Added `Check` and `ListCheck` support to existing classes
   - added `IsStringy::check()`
   - `is_stringy()` now uses `IsStringy::check()`